fix:solved format in  product item card and seed db
