+++ app/(public)/(home)/_components/top-deal-items.tsx
@@ -1,10 +1,9 @@
 'use client';
+import { ProductItemSm } from '@/app/(public)/_components/product-item-sm';
 import { productItemType } from '@/types/public.data-types';
 import { useTranslations } from 'next-intl';
 import Link from 'next/link';
 import { AiFillLike } from 'react-icons/ai';
-import { ProductItem } from '../../_components/product-item';
-import { ProductItemSm } from '@/app/(public)/_components/product-item-sm';
 
 type TopDealItemsProps = {
   data: productItemType[];
@@ -31,12 +30,12 @@ export const TopDealItems = ({
         </Link>
       </div>
       {/* item list */}
-      <div className="w-full flex flex-row justify-between p-2 gap-3 overflow-x-auto">
+      <div
+        className={`w-full grid grid-cols-${size} p-2 gap-3 overflow-x-auto`}
+      >
         {data.map((item: productItemType, index) => (
           <ProductItemSm
-            key={item.id}
             item={item}
-            size={size}
             renderSaleValue={renderSaleValue}
           />
         ))}
