+++ app/(public)/_components/product-item.tsx
@@ -1,18 +1,17 @@
-import { Separator } from '@/components/ui/separator';
-import { productItemType } from '@/types/public.data-types';
-import { useRouter } from 'next/navigation';
-import { formatPrice } from './global-function';
-import { RatingStars } from './rating-starts';
-import Image from 'next/image';
 import {
   Card,
-  CardAction,
   CardContent,
   CardDescription,
   CardFooter,
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
+import { Separator } from '@/components/ui/separator';
+import { productItemType } from '@/types/public.data-types';
+import Image from 'next/image';
+import { useRouter } from 'next/navigation';
+import { formatPrice } from './global-function';
+import { RatingStars } from './rating-starts';
 
 interface ProductItemProps {
   item: productItemType;
@@ -29,53 +28,65 @@ export const ProductItem = ({
   const handleOpenDetail = (id: string) => {
     router.push(`/products/${id}`);
   };
-  const renderSaleInfo = (type: string | null, value: number | null) => {
+  const renderSaleInfo = ({
+    voucher,
+  }: {
+    voucher: { type: string; value: number; maxDiscount: number } | null;
+  }) => {
+    console.log(voucher);
     const originalPrice = (
       <span className="text-[var(--muted-foreground)] line-through text-xs">
-        {formatPrice(item.minPrice)}đ
+        {formatPrice(item.minPrice)}
       </span>
     );
 
-    if (!type || value === null || !renderSaleValue) {
+    if (!voucher) {
       return (
         <div className="flex items-center gap-2">
-          <span className="font-medium">{formatPrice(item.minPrice)}đ</span>
+          <p className="font-medium">{formatPrice(item.minPrice)}</p>
+        </div>
+      );
+    } else if (voucher) {
+      const discountedPrice =
+        voucher.type === 'PERCENT'
+          ? item.minPrice - (item.minPrice * Number(voucher.value)) / 100
+          : item.minPrice - Number(voucher.value);
+
+      const promotionBadge =
+        voucher.type === 'PERCENT' ? (
+          <span className="bg-green-100 text-green-600 text-xs font-medium px-2 py-[2px] rounded-md">
+            -{voucher.value}%
+          </span>
+        ) : (
+          <span className="bg-green-100 text-green-600 text-xs font-medium px-2 py-[2px] rounded-md">
+            -{formatPrice(Number(voucher.value))}
+          </span>
+        );
+
+      return (
+        <div className="flex flex-col items-start gap-2">
+          <div className="flex flex-col justify-start items-start gap-0.5">
+            <div className="text-[var(--destructive)] font-medium text-[16px]">
+              {Number(discountedPrice) <= 0 ? (
+                <div className="flex flex-row  gap-1">
+                  0<div className="underline">{'đ'}</div>
+                </div>
+              ) : (
+                formatPrice(discountedPrice)
+              )}
+            </div>
+            <div>{originalPrice}</div>
+          </div>
+          {promotionBadge}
         </div>
       );
     }
-
-    const discountedPrice =
-      type === 'PERCENT'
-        ? item.minPrice - (item.minPrice * value) / 100
-        : item.minPrice - value;
-
-    const promotionBadge =
-      type === 'PERCENT' ? (
-        <span className="bg-green-100 text-green-600 text-xs font-medium px-2 py-[2px] rounded-md">
-          -{value}%
-        </span>
-      ) : (
-        <span className="bg-green-100 text-green-600 text-xs font-medium px-2 py-[2px] rounded-md">
-          -{formatPrice(value)}
-        </span>
-      );
-
-    return (
-      <div className="flex flex-col items-start gap-2">
-        <div className="flex flex-row items-center gap-2">
-          <p className="text-[var(--destructive)] font-medium text-sm">
-            {formatPrice(discountedPrice)}
-          </p>
-          {originalPrice}
-        </div>
-        {promotionBadge}
-      </div>
-    );
   };
 
   return (
     <div
-      className={`w-${size} flex flex-col justify-start items-start flex-${size} border border-gray-200 rounded-lg hover:cursor-pointer`}
+      key={item.id}
+      className={`w-full flex flex-col justify-start items-start flex-1 border border-gray-200 rounded-lg hover:cursor-pointer`}
       onClick={() => handleOpenDetail(item.id)}
     >
       <div className="relative w-full flex-shrink-0 aspect-square overflow-hidden rounded-t-lg">
@@ -86,10 +97,10 @@ export const ProductItem = ({
           className="object-cover"
         />
       </div>
-      <Card className="w-full max-w-sm shadow-none border-none rounded-t-none flex-shrink-0">
-        <CardHeader>
+      <Card className="w-full max-w-sm shadow-none border-none rounded-t-none flex-shrink-0 py-2">
+        <CardHeader className="px-2">
           <CardTitle className="overflow-hidden">
-            <p className="text-base font-normal line-clamp-1">{item.title}</p>
+            <p className="text-base font-normal line-clamp-2">{item.title}</p>
           </CardTitle>
           <CardDescription>
             <div className="w-full">
@@ -98,18 +109,18 @@ export const ProductItem = ({
                 quắc quắc quặc qucặ
               </p>
             </div>
-            <RatingStars value={item.ratingAvg} size={15}/>
+            <RatingStars value={item.ratingAvg} size={15} />
           </CardDescription>
         </CardHeader>
-        <CardContent className="flex flex-col justify-start items-start">
+        <CardContent className="flex flex-col justify-start items-start px-2">
           {/*{renderSalePrice(item.voucher.type, item.voucher.value)}*/}
           {/*{renderPromotionType(item.voucher.type)}*/}
-          {renderSaleInfo(item.voucher.type, item.voucher.value)}
+          {renderSaleInfo({ voucher: item.voucher })}
         </CardContent>
-        <CardFooter className="flex-col flex justify-center items-start">
+        <CardFooter className="flex-col flex justify-center items-start px-2 py-1">
           <Separator />
-          <p className="p-2 text-[var(--muted-foreground)] text-xs">
-            Made in {'VN'}
+          <p className="w-full p-2 text-[var(--muted-foreground)] text-xs line-clamp-1 whitespace-nowrap overflow-x-hidden overflow-ellipsis">
+            Made in {item.origin}
           </p>
         </CardFooter>
       </Card>
