+++ app/(public)/products/[id]/_components/suggest-deal-today.tsx
@@ -16,9 +16,11 @@ export const SuggestDealToday = ({ data }: TopDealItemsProps) => {
       {/* top-title */}
       <p className="w-full text-left font-bold">{t('title')}</p>
       {/* content here */}
-      <div className="grid grid-cols-5 gap-2 w-full">
+      <div className="grid grid-cols-4 gap-2 w-full">
         {data.map((item, index) => (
-          <ProductItem key={index} item={item} size="1" />
+          <div key={index} className='w-full'>
+            <ProductItem item={item} size="1" />
+          </div>
         ))}
       </div>
       {/* watch more */}
